+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@nevermined-io/payments",
-  "version": "0.7.0",
+  "version": "0.7.1",
   "description": "Typescript SDK to interact with the Nevermined Payments Protocol",
   "main": "./dist/index.js",
   "types": "./dist/index.d.ts",
